requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Google Gemini API
google-generativeai>=0.3.0

# Vector Database
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0

# Additional Dependencies
python-dotenv==1.0.0
requests>=2.31.0
numpy>=1.26.0
Jinja2==3.1.2

# Production Server (optional)
gunicorn==21.2.0

# Development Tools (optional)
pytest>=7.4.0
black>=23.12.0

